Code restructuring, moving more widely used code into base db package.
