+++ back-end/db/db.go
@@ -1,37 +1,13 @@
-package main
+package db
 
 import (
 	"fmt"
-	"log"
 	"os"
 	"strings"
-
-	"../../config"
-	"../schema"
-	"../seeds"
-
-	"github.com/jinzhu/gorm"
-	_ "github.com/jinzhu/gorm/dialects/postgres"
-	_ "github.com/lib/pq"
 )
 
-func main() {
-	config.RetrieveDBEnvVariables()
-
-	dbDialogue := createConnectionDialogue()
-
-	db, err := gorm.Open("postgres", dbDialogue)
-	if err != nil {
-		panic(err)
-	}
-	defer db.Close()
-	log.Println("Connection Established")
-
-	schema.Migrate(db)
-	seeds.Seed(db)
-}
-
-func createConnectionDialogue() string {
+// CreateConnectionDialogue creates string to connect to pgress db.
+func CreateConnectionDialogue() string {
 	var dbDialogue strings.Builder
 
 	dbDialogue.WriteString("host=")
