+++ back-end/go.mod
@@ -3,7 +3,10 @@ module github.com/alumni-lab/sap-clone
 go 1.13
 
 require (
+	github.com/DATA-DOG/go-sqlmock v1.4.1
 	github.com/gin-gonic/gin v1.5.0
 	github.com/jinzhu/gorm v1.9.12
 	github.com/lib/pq v1.1.1
+	github.com/spf13/viper v1.6.2
+	github.com/stretchr/testify v1.4.0
 )
